Use abs_path like error stacktrace frames
