+++ src/js/profiling/convertHermesProfile.ts
@@ -2,7 +2,7 @@ import type { FrameId, StackId, ThreadCpuFrame, ThreadCpuSample, ThreadCpuStack,
 import { logger } from '@sentry/utils';
 
 import type * as Hermes from './hermes';
-import { parseHermesJSStackFrame } from './hermes';
+import { DEFAULT_BUNDLE_NAME } from './hermes';
 import { MAX_PROFILE_DURATION_MS } from './integration';
 import type { RawThreadCpuProfile } from './types';
 
@@ -171,5 +171,43 @@ function mapStacks(
   return {
     stacks,
     hermesStackToSentryStackMap,
+  };
+}
+
+/**
+ * Parses Hermes StackFrame to Sentry StackFrame.
+ * For native frames only function name is returned, for Hermes bytecode the line and column are calculated.
+ */
+export function parseHermesJSStackFrame(frame: Hermes.StackFrame): ThreadCpuFrame {
+  if (frame.category !== 'JavaScript') {
+    // Native
+    if (frame.name === '[root]') {
+      return { function: frame.name, in_app: false };
+    }
+    return { function: frame.name };
+  }
+
+  if (frame.funcVirtAddr !== undefined && frame.offset !== undefined) {
+    // Hermes Bytecode
+    return {
+      function: frame.name,
+      abs_path: DEFAULT_BUNDLE_NAME,
+      // https://github.com/krystofwoldrich/metro/blob/417e6f276ff9422af6039fc4d1bce41fcf7d9f46/packages/metro-symbolicate/src/Symbolication.js#L298-L301
+      // Hermes lineno is hardcoded 1, currently only one bundle symbolication is supported by metro-symbolicate and thus by us.
+      lineno: 1,
+      // Hermes colno is 0-based, while Sentry is 1-based
+      colno: Number(frame.funcVirtAddr) + Number(frame.offset) + 1,
+    };
+  }
+
+  // JavaScript
+  const indexOfLeftParenthesis = frame.name.indexOf('(');
+  return {
+    function:
+      indexOfLeftParenthesis !== -1 && frame.name.substring(0, indexOfLeftParenthesis) ||
+      frame.name,
+    abs_path: DEFAULT_BUNDLE_NAME,
+    lineno: frame.line !== undefined ? Number(frame.line) : undefined,
+    colno: frame.column !== undefined ? Number(frame.column) : undefined,
   };
 }
