+++ src/js/profiling/hermes.ts
@@ -53,55 +53,8 @@ export interface Profile {
   stackFrames: Record<string, StackFrame>;
 }
 
-export interface ParsedHermesStackFrame {
-  function: string;
-  file?: string;
-  lineno?: number;
-  colno?: number;
-  in_app?: boolean;
-}
-
 export const DEFAULT_BUNDLE_NAME =
   Platform.OS === 'android' ? ANDROID_DEFAULT_BUNDLE_NAME : Platform.OS === 'ios' ? IOS_DEFAULT_BUNDLE_NAME : undefined;
-const ANONYMOUS_FUNCTION_NAME = 'anonymous';
-
-/**
- * Parses Hermes StackFrame to Sentry StackFrame.
- * For native frames only function name is returned, for Hermes bytecode the line and column are calculated.
- */
-export function parseHermesJSStackFrame(frame: StackFrame): ParsedHermesStackFrame {
-  if (frame.category !== 'JavaScript') {
-    // Native
-    if (frame.name === '[root]') {
-      return { function: frame.name, in_app: false };
-    }
-    return { function: frame.name };
-  }
-
-  if (frame.funcVirtAddr !== undefined && frame.offset !== undefined) {
-    // Hermes Bytecode
-    return {
-      function: frame.name || ANONYMOUS_FUNCTION_NAME,
-      file: DEFAULT_BUNDLE_NAME,
-      // https://github.com/krystofwoldrich/metro/blob/417e6f276ff9422af6039fc4d1bce41fcf7d9f46/packages/metro-symbolicate/src/Symbolication.js#L298-L301
-      // Hermes lineno is hardcoded 1, currently only one bundle symbolication is supported by metro-symbolicate and thus by us.
-      lineno: 1,
-      // Hermes colno is 0-based, while Sentry is 1-based
-      colno: Number(frame.funcVirtAddr) + Number(frame.offset) + 1,
-    };
-  }
-
-  // JavaScript
-  const indexOfLeftParenthesis = frame.name.indexOf('(');
-  return {
-    function:
-      (indexOfLeftParenthesis !== -1 && (frame.name.substring(0, indexOfLeftParenthesis) || ANONYMOUS_FUNCTION_NAME)) ||
-      frame.name,
-    file: DEFAULT_BUNDLE_NAME,
-    lineno: frame.line !== undefined ? Number(frame.line) : undefined,
-    colno: frame.column !== undefined ? Number(frame.column) : undefined,
-  };
-}
 
 const MS_TO_NS: number = 1e6;
